enahnce:1.修复选择排序的排序问题(重新设计字段的类型解决的)2.计算距离不重置0 3.增加把数据库移到/storage/sdcard0/trace.db不用root也可以拿到出来(不是每次都拿，需要拿的时候去掉注释重新编译一次)

+++ app/src/main/java/com/example/fazhao/locationmanager/baidu_map/model/TraceItem.java
@@ -11,11 +11,11 @@ public class TraceItem {
     private int tag;
     private int step = -1;
 
-    public int getUptime() {
+    public long getUptime() {
         return uptime;
     }
 
-    public void setUptime(int uptime) {
+    public void setUptime(long uptime) {
         this.uptime = uptime;
     }
 
@@ -27,7 +27,7 @@ public class TraceItem {
         this.distance = distance;
     }
 
-    private int uptime = -1;
+    private long uptime = -1;
     private double distance = -1;
 
     public int getStep() {
